+++ kernel/memory/virtual/vmm/vmm.h
@@ -5,7 +5,7 @@
 
 extern Bitmap vmm_bitmap;
 
-uint64_t vmm_malloc_pages(const uint64_t byteAmount);
-bool vmm_free_pages(const uint64_t virtAddr, const uint64_t byteAmount);
+[[nodiscard]] uint64_t vmm_malloc_pages(const uint64_t byteAmount);
+[[nodiscard]] bool vmm_free_pages(const uint64_t virtAddr, const uint64_t byteAmount);
 
-bool init_vmm();
+[[nodiscard]] bool init_vmm();